src/test/java/com/decagon/fintechpaymentapisqd11b/service/serviceImpl/LoginServiceImplTest.java
package com.decagon.fintechpaymentapisqd11b.service.serviceImpl;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.mockito.Mockito.any;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

import com.decagon.fintechpaymentapisqd11b.dto.LoginRequestPayload;
import com.decagon.fintechpaymentapisqd11b.security.filter.JwtUtils;

import java.util.ArrayList;

import org.junit.jupiter.api.Disabled;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.BadCredentialsException;
import org.springframework.security.authentication.TestingAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.AuthenticationException;
import org.springframework.security.core.userdetails.User;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit.jupiter.SpringExtension;

@ContextConfiguration(classes = {LoginServiceImpl.class})
@ExtendWith(SpringExtension.class)
class LoginServiceImplTest {
    @MockBean
    private AuthenticationManager authenticationManager;

    @MockBean
    private JwtUtils jwtUtils;

    @Autowired
    private LoginServiceImpl loginServiceImpl;

    @MockBean
    private UsersServiceImpl usersServiceImpl;


    @Test
    void testLogin() throws AuthenticationException {
        when(usersServiceImpl.loadUserByUsername((String) any()))
                .thenReturn(new User("Lawal", "12345", new ArrayList<>()));
        when(jwtUtils.generateToken((UserDetails) any())).thenReturn("login successful");
        when(authenticationManager.authenticate((Authentication) any()))
                .thenReturn(new TestingAuthenticationToken("Principal", "Credentials"));
        assertEquals("login successful", loginServiceImpl.login(new LoginRequestPayload("lawalmonsaw@gmail.com", "12345")));
        verify(usersServiceImpl).loadUserByUsername((String) any());
        verify(jwtUtils).generateToken((UserDetails) any());
        verify(authenticationManager).authenticate((Authentication) any());
    }



}

